Implementacion de la funcion para control de medicamentos,nuevos recursos para la UI y mejoras en la UI principal(dashboard)

+++ app/src/main/java/com/joseantonio/personalproject/proyectovighealth/consultasDb/ConsultasPesoImpl.java
@@ -118,7 +118,7 @@ public class ConsultasPesoImpl extends DbHelper implements ConsultasPeso {
 
         if(cursorPeso.moveToFirst()){
             do{
-                peso = new Peso(cursorPeso.getDouble(0), cursorPeso.getFloat(1));
+                peso = new Peso(cursorPeso.getDouble(0), cursorPeso.getDouble(1));
 
             }while(cursorPeso.moveToNext());
         }
